Moved RIDFlight, AircraftState and AircraftHeight to rid module, updated imports and test payload generator
